src/backend/pyproject.toml
[tool.black]
line-length = 120
include = '\.pyi?$'
skip-string-normalization = 'true'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
source = "."

[tool.pdm]
distribution = false

[tool.pdm.resolution]
respect-source-order = true

[project]
name = "bk-audit"
version = "1.15.0"
description = "Default template for PDM package"
dependencies = [
    "raven==6.10.0",
    "sentry-sdk==1.45.1",
    "bkapi-component-open>=1.1.0",
    "bk-audit[opentelemetry]==1.2.2b0",
    "bk-resource==0.4.11",
    "blueapps[opentelemetry]==4.16rc2",
    "gunicorn==23.0.0",
    "django~=4.2.26",
    "django-filter==2.4.0",
    "django-cors-headers==4.7.0",
    "django-cors-middleware==1.3.1",
    "djangorestframework~=3.15.2",
    "django-rest-framework-condition==0.1.1",
    "drf-yasg==1.21.9",
    "drf-spectacular==0.28.0",
    "django-environ==0.10.0",
    "django-simpleui==2023.3.1",
    "django-sslserver==0.22",
    "pymysql==1.1.1",
    "whitenoise==5.2.0",
    "redis==3.5.3",
    "django-redis==5.2.0",
    "celery==5.5.3",
    "django-celery-beat==2.7.0",
    "django-celery-results==2.4.0",
    "bk-iam==1.3.2",
    "ipython==8.10.0",
    "pyinstrument==3.4.2",
    "JSON-log-formatter==0.4.0",
    "arrow==1.2.2",
    "requests==2.32.4",
    "mistune==2.0.3",
    "protobuf<6.0",
    "opentelemetry-api==1.30.0",
    "opentelemetry-sdk==1.30.0",
    "opentelemetry-exporter-otlp==1.30.0",
    "opentelemetry-exporter-jaeger==1.21.0",
    "opentelemetry-instrumentation==0.51b0",
    "opentelemetry-instrumentation-django==0.51b0",
    "opentelemetry-instrumentation-dbapi==0.51b0",
    "opentelemetry-instrumentation-redis==0.51b0",
    "opentelemetry-instrumentation-requests==0.51b0",
    "opentelemetry-instrumentation-celery==0.51b0",
    "opentelemetry-instrumentation-logging==0.51b0",
    "django-prometheus==2.2.0",
    "numpy==1.24.2",
    "apigw-manager==4.2.0",
    "openpyxl==3.1.2",
    "elasticsearch==7.17.9",
    "pypinyin==0.49.0",
    "bk-crypto-python-sdk==2.0.1",
    "client-throttler==2.1.0",
    "bk-notice-sdk==1.3.2",
    "pypika==0.48.9",
    "pydantic==2.10.4",
    "kafka-python==2.0.2",
    "stackprinter==0.2.12",
    "jinja2>=3.1.5",
    "opentelemetry-instrumentation-kafka-python==0.51b0",
    "opentelemetry-instrumentation-urllib3==0.51b0",
    "opentelemetry-instrumentation-httpx==0.51b0",
    "opentelemetry-instrumentation-asyncio==0.51b0",
    "opentelemetry-instrumentation-threading==0.51b0",
    "django-treebeard==4.7.1",
    "bkstorages==2.0.0",
    "gevent==25.4.1",
    "xlsxwriter==3.2.3",
    "nh3>=0.3.0",
    "flower==2.0.1",
    "pympler>=1.1",
    "drf-pydantic==2.7.1",
    "sqlglot==26.26.0",
    "django-db-connection-pool==1.2.6",
    "supervisor~=4.3.0"
]
requires-python = "==3.11.*"
readme = "README.md"
license = { text = "MIT" }

[dependency-groups]
dev = [
    "pre-commit==4.2.0",
    "flake8==3.7.8",
    "seed-isort-config==1.9.2",
    "coverage==7.8.2",
    "pytest==8.3.4 ",
    "pytest-cov==3.0.0",
    "pytest-django==4.10.0",
    "pytest-xdist==2.4.0",
    "pytest-html==3.1.1",
    "xmlrunner==1.7.7",
    "pyparsing==2.2.0",
    "PyYAML==6.0",
    "packaging==24.2",
]
